+++ src/ignitus-Dashboard/ignitus-Student-Dashboard/Components/Dashboard.js
@@ -4,7 +4,7 @@
 import React from 'react';
 import '../Styles/style.scss';
 import { withErrorBoundary } from '../../../ignitus-Shared/Components/errorBoundary';
-import SideNavigation from '../../../ignitus-Shared/Components/emotionStyles/ignitus-Templates/DashboardSideNavigation/SideNavigation';
+import SideNavigation from '../../../ignitus-Shared/Components/ignitus-DesignSystem/ignitus-Templates/DashboardSideNavigation/SideNavigation';
 
 const Dashboard = () => {
   let email = '';
