+++ src/ignitus-Shared/Components/ignitus-DesignSystem/ignitus-Templates/DashboardSideNavigation/style.scss
